device_control/requirements.txt
fastapi
uvicorn
paho-mqtt
pymongo
python-multipart
